requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
email_validator
